+++ co-simulation/fed/mosaic-carma-messenger/pom.xml
@@ -109,6 +109,12 @@
         <version>22.1-SNAPSHOT</version>
         <scope>compile</scope>
       </dependency>
+      <dependency>
+        <groupId>org.eclipse.mosaic</groupId>
+        <artifactId>mosaic-common-utils</artifactId>
+        <version>22.1-SNAPSHOT</version>
+        <scope>compile</scope>
+      </dependency>
   </dependencies>
 
   <profiles>
