+++ co-simulation/lib/mosaic-common-utils/src/main/java/org/eclipse/mosaic/lib/CommonUtil/CommonInstance.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.eclipse.mosaic.lib.util;
+package org.eclipse.mosaic.lib.CommonUtil;
 
 import java.io.IOException;
 import java.net.DatagramPacket;
