+++ co-simulation/lib/mosaic-common-utils/src/test/java/org/eclipse/mosaic/lib/CommonUtil/CommonInstanceManagerTest.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.eclipse.mosaic.lib.util;
+package org.eclipse.mosaic.lib.CommonUtil;
 
 import java.io.IOException;
 import java.net.InetAddress;
