+++ co-simulation/lib/mosaic-common-utils/src/test/java/org/eclipse/mosaic/lib/CommonUtil/CommonMessageAmbassadorTest.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package org.eclipse.mosaic.lib.util;
+package org.eclipse.mosaic.lib.CommonUtil;
 
 import java.io.File;
 import java.io.IOException;
